// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const VerificationStatus = {
  PENDING: 'PENDING',
  IN_PROGRESS: 'IN_PROGRESS',
  CLEAR: 'CLEAR',
  DISCREPANCY: 'DISCREPANCY',
  FAILED: 'FAILED'
} as const

export type VerificationStatus = (typeof VerificationStatus)[keyof typeof VerificationStatus]


export const DocumentType = {
  OFFER_LETTER: 'OFFER_LETTER',
  RELIEVING_LETTER: 'RELIEVING_LETTER'
} as const

export type DocumentType = (typeof DocumentType)[keyof typeof DocumentType]